Add empty folder indicator
Whenever a folder is empty in any of the files view, an icon and a
message is displayed. This was done by extending the generic FileList
component with an additional slot for empty folder messages.

Also, the header bar is removed when no files are visible.

In the trashbin view, the button "Clear trashbin" is now also disabled
when there are no files to delete.